chore: rename server to peer
