+++ go-peer/.gitignore
@@ -1,0 +1,1 @@
+identity.key
