--- go-server/.gitignore
@@ -1,1 +1,0 @@
-identity.key
