--- go-server/Dockerfile
@@ -1,12 +1,0 @@
-FROM golang:1.20-alpine
-
-WORKDIR /usr/src/app
-
-# pre-copy/cache go.mod for pre-downloading dependencies and only redownloading them in subsequent builds if they change
-COPY go.mod go.sum ./
-RUN go mod download && go mod verify
-
-COPY . .
-RUN go build -v -o /usr/local/bin/universal-chat-go ./...
-
-CMD ["universal-chat-go"]
