--- go-server/identity.go
@@ -1,49 +1,0 @@
-package main
-// Borrowed from https://github.com/libp2p/go-libp2p-relay-daemon/blob/master/identity.go
-
-import (
-	"fmt"
-	"os"
-
-	"github.com/libp2p/go-libp2p/core/crypto"
-)
-
-// LoadIdentity reads a private key from the given path and, if it does not
-// exist, generates a new one.
-func LoadIdentity(idPath string) (crypto.PrivKey, error) {
-	if _, err := os.Stat(idPath); err == nil {
-		return ReadIdentity(idPath)
-	} else if os.IsNotExist(err) {
-		fmt.Printf("Generating peer identity in %s\n", idPath)
-		return GenerateIdentity(idPath)
-	} else {
-		return nil, err
-	}
-}
-
-// ReadIdentity reads a private key from the given path.
-func ReadIdentity(path string) (crypto.PrivKey, error) {
-	bytes, err := os.ReadFile(path)
-	if err != nil {
-		return nil, err
-	}
-
-	return crypto.UnmarshalPrivateKey(bytes)
-}
-
-// GenerateIdentity writes a new random private key to the given path.
-func GenerateIdentity(path string) (crypto.PrivKey, error) {
-	privk, _, err := crypto.GenerateKeyPair(crypto.Ed25519, 0)
-	if err != nil {
-		return nil, err
-	}
-
-	bytes, err := crypto.MarshalPrivateKey(privk)
-	if err != nil {
-		return nil, err
-	}
-
-	err = os.WriteFile(path, bytes, 0400)
-
-	return privk, err
-}
