+++ rust-peer/.dockerignore
@@ -1,0 +1,1 @@
+target/
