+++ rust-peer/Cargo.toml
@@ -1,0 +1,33 @@
+[package]
+name = "rust-libp2p-webrtc-peer"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+anyhow = "1.0"
+clap = { version = "4.1.11", features = ["derive"] }
+env_logger = "0.10.0"
+futures = "0.3.27"
+futures-timer = "3.0.2"
+libp2p = { git = "https://github.com/vnermolaev/rust-libp2p.git", branch = "deprecate/gossipsub-close-event", version = "0.51.2", features = [
+	"identify",
+	"ping",
+	"tokio",
+	"gossipsub",
+	"webrtc",
+	"macros",
+	"relay",
+	"kad",
+	"rsa",
+	"ed25519",
+] }
+libp2p-webrtc = { git = "https://github.com/vnermolaev/rust-libp2p.git", branch = "deprecate/gossipsub-close-event", version = "0.4.0-alpha.4", features = [
+	"tokio",
+] }
+log = "0.4.17"
+rand = "0.8.5"
+tokio = { version = "1.26.0", features = ["full"] }
+tokio-util = { version = "0.7", features = ["full"] }
+webrtc = { version = "0.6.0", optional = true }
