--- rust-server/.dockerignore
@@ -1,1 +1,0 @@
-target/
