--- rust-server/Dockerfile
@@ -1,9 +1,0 @@
-FROM rust:1.68 as builder
-WORKDIR /usr/src/universal-chat
-COPY . .
-RUN cargo install --path .
-
-FROM debian:bullseye-slim
-RUN apt-get update && apt-get install && rm -rf /var/lib/apt/lists/*
-COPY --from=builder /usr/local/cargo/bin/rust-libp2p-webrtc-server /usr/local/bin/rust-libp2p-webrtc-server
-CMD ["rust-libp2p-webrtc-server"]
