--- rust-server/src/main.rs
@@ -1,195 +1,0 @@
-use anyhow::Result;
-use clap::Parser;
-use futures::StreamExt;
-use libp2p::{
-    core::muxing::StreamMuxerBox,
-    gossipsub, identify, identity,
-    kad::record::store::MemoryStore,
-    kad::{Kademlia, KademliaConfig},
-    multiaddr::Protocol,
-    ping, relay,
-    swarm::{keep_alive, NetworkBehaviour, Swarm, SwarmBuilder, SwarmEvent},
-    Multiaddr, PeerId, Transport,
-};
-use libp2p_webrtc as webrtc;
-use log::{debug, error, info, warn};
-use rand::thread_rng;
-use std::hash::{Hash, Hasher};
-use std::time::Instant;
-use std::{collections::hash_map::DefaultHasher, time::Duration};
-
-// TODO: replace with our private bootstrap node
-// const BOOTNODES: [&str; 4] = [
-//     "QmNnooDu7bfjPFoTZYxMNLWUQJyrVwtbZg5gBMjTezGAJN",
-//     "QmQCU2EcMqAqQPR2i9bChDtGNJchTbq5TbXJJ16u19uLTa",
-//     "QmbLHAnMoJPWSCR5Zhtx6BHJX9KiKNN6tpvbUcqanj75Nb",
-//     "QmcZf59bWwK5XFi76CZX8cbJ4BhTzzA3gU1ZjYZcYW3dwt",
-// ];
-
-const TICK_INTERVAL: Duration = Duration::from_secs(5);
-
-#[derive(Debug, Parser)]
-#[clap(name = "universal connectivity rust server")]
-struct Opt {
-    /// Address of a remote peer to connect to.
-    #[clap(long)]
-    remote_address: Option<Multiaddr>,
-}
-
-/// An example WebRTC server that will accept connections
-#[tokio::main]
-async fn main() -> Result<()> {
-    env_logger::Builder::from_env(env_logger::Env::default().default_filter_or("info")).init();
-
-    let opt = Opt::parse();
-
-    let mut swarm = create_swarm()?;
-    swarm.listen_on(format!("/ip4/127.0.0.1/udp/0/webrtc").parse()?)?;
-
-    if let Some(remote_address) = opt.remote_address {
-        swarm.dial(remote_address).unwrap();
-    }
-
-    let mut tick = futures_timer::Delay::new(TICK_INTERVAL);
-
-    let now = Instant::now();
-    loop {
-        match futures::future::select(swarm.next(), &mut tick).await {
-            futures::future::Either::Left((event, _)) => match event.unwrap() {
-                SwarmEvent::NewListenAddr { address, .. } => {
-                    let p2p_address = address.with(Protocol::P2p((*swarm.local_peer_id()).into()));
-                    info!("Listen address: {p2p_address:?}")
-                }
-                SwarmEvent::ConnectionEstablished { peer_id, .. } => {
-                    info!("Connected to {peer_id}");
-                }
-                SwarmEvent::OutgoingConnectionError { peer_id, error } => {
-                    warn!("Failed to dial {peer_id:?}: {error}");
-                }
-                SwarmEvent::ConnectionClosed { peer_id, cause, .. } => {
-                    warn!("Connection to {peer_id} closed: {cause:?}");
-                }
-                SwarmEvent::Behaviour(BehaviourEvent::Relay(e)) => {
-                    info!("{:?}", e);
-                }
-                SwarmEvent::Behaviour(BehaviourEvent::Gossipsub(
-                    libp2p::gossipsub::Event::Message {
-                        message_id: _,
-                        propagation_source: _,
-                        message,
-                    },
-                )) => {
-                    info!(
-                        "Received message from {:?}: {}",
-                        message.source,
-                        String::from_utf8(message.data).unwrap()
-                    );
-                }
-                SwarmEvent::Behaviour(BehaviourEvent::Gossipsub(
-                    libp2p::gossipsub::Event::Subscribed { peer_id, topic },
-                )) => {
-                    info!("{peer_id} subscribed to {topic}");
-                }
-                event => {
-                    debug!("{event:?}");
-                }
-            },
-            futures::future::Either::Right(_) => {
-                tick = futures_timer::Delay::new(TICK_INTERVAL);
-
-                let message = format!("Hello world! Sent at: {:4}s", now.elapsed().as_secs_f64());
-
-                if let Err(err) = swarm.behaviour_mut().gossipsub.publish(
-                    gossipsub::IdentTopic::new("universal-connectivity"),
-                    message.as_bytes(),
-                ) {
-                    error!("Failed to publish periodic message: {err}")
-                }
-            }
-        }
-    }
-}
-
-#[derive(NetworkBehaviour)]
-struct Behaviour {
-    gossipsub: gossipsub::Behaviour,
-    identify: identify::Behaviour,
-    kademlia: Kademlia<MemoryStore>,
-    keep_alive: keep_alive::Behaviour,
-    ping: ping::Behaviour,
-    relay: relay::Behaviour,
-}
-
-fn create_swarm() -> Result<Swarm<Behaviour>> {
-    let local_key = identity::Keypair::generate_ed25519();
-    let local_peer_id = PeerId::from(local_key.public());
-    debug!("Local peer id: {local_peer_id}");
-
-    // To content-address message, we can take the hash of message and use it as an ID.
-    let message_id_fn = |message: &gossipsub::Message| {
-        let mut s = DefaultHasher::new();
-        message.data.hash(&mut s);
-        gossipsub::MessageId::from(s.finish().to_string())
-    };
-
-    // Set a custom gossipsub configuration
-    let gossipsub_config = gossipsub::ConfigBuilder::default()
-        .validation_mode(gossipsub::ValidationMode::Permissive) // This sets the kind of message validation. The default is Strict (enforce message signing)
-        .message_id_fn(message_id_fn) // content-address messages. No two messages of the same content will be propagated.
-        .mesh_outbound_min(1)
-        .mesh_n_low(1)
-        .flood_publish(true)
-        .build()
-        .expect("Valid config");
-
-    // build a gossipsub network behaviour
-    let mut gossipsub = gossipsub::Behaviour::new(
-        gossipsub::MessageAuthenticity::Signed(local_key.clone()),
-        gossipsub_config,
-    )
-    .expect("Correct configuration");
-
-    // Create a Gossipsub topic
-    let topic = gossipsub::IdentTopic::new("universal-connectivity");
-
-    // subscribes to our topic
-    gossipsub.subscribe(&topic)?;
-
-    let transport = webrtc::tokio::Transport::new(
-        local_key.clone(),
-        webrtc::tokio::Certificate::generate(&mut thread_rng())?,
-    );
-
-    let identify_config = identify::Behaviour::new(
-        identify::Config::new("/ipfs/0.1.0".into(), local_key.public().clone())
-            .with_initial_delay(Duration::ZERO),
-    );
-
-    // Create a Kademlia behaviour.
-    let mut cfg = KademliaConfig::default();
-    cfg.set_query_timeout(Duration::from_secs(5 * 60));
-    let store = MemoryStore::new(local_peer_id);
-    let kad_behaviour = Kademlia::with_config(local_peer_id, store, cfg);
-
-    // Add the bootnodes to the local routing table. `libp2p-dns` built
-    // into the `transport` resolves the `dnsaddr` when Kademlia tries
-    // to dial these nodes.
-    // for peer in &BOOTNODES {
-    //     // TODO: update this
-    //     kad_behaviour.add_address(&peer.parse()?, "/dnsaddr/bootstrap.libp2p.io".parse()?);
-    // }
-
-    let transport = transport
-        .map(|(local_peer_id, conn), _| (local_peer_id, StreamMuxerBox::new(conn)))
-        .boxed();
-
-    let behaviour = Behaviour {
-        gossipsub,
-        identify: identify_config,
-        kademlia: kad_behaviour,
-        keep_alive: keep_alive::Behaviour::default(),
-        ping: ping::Behaviour::default(),
-        relay: relay::Behaviour::new(local_peer_id, Default::default()),
-    };
-    Ok(SwarmBuilder::with_tokio_executor(transport, behaviour, local_peer_id).build())
-}
